Extract period (seasonal) metrics from fitted GAM

Extract period (seasonal) metrics from fitted GAM

anlz_metseason(mod, metfun = mean, doystr = 1, doyend = 364, nsim = 10000, ...)

Arguments

mod
input model object as returned by anlz_gam

metfun
function input for metric to calculate, e.g., mean, var, max, etc

doystr
numeric indicating start Julian day for extracting averages

doyend
numeric indicating ending Julian day for extracting averages

nsim
numeric indicating number of random draws for simulating uncertainty

...
additional arguments passed to metfun, e.g., na.rm = TRUE)

Value

A data frame of period metrics

Details

This function estimates a metric of interest for a given seasonal period each year using results from a fitted GAM (i.e., from anlz_gam). The estimates are based on the predicted values for each seasonal period, with uncertainty of the metric based on repeated sampling of the predictions following uncertainty in the model coefficients.

Examples

library(dplyr)

# data to model
tomod <- rawdat %>%
  filter(station %in% 34) %>%
  filter(param %in% 'chl')

mod <- anlz_gam(tomod, trans = 'log10')
anlz_metseason(mod, mean, doystr = 90, doyend = 180, nsim = 100)

#> # A tibble: 29 x 6
#>       yr   met     se bt_lwr bt_upr bt_met
#>    <dbl> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1  1991 0.113 0.527   0.135  15.7    1.46
#>  2  1992 0.684 0.0615  4.10    7.15   5.41
#>  3  1993 0.689 0.0969  3.54    8.48   5.48
#>  4  1994 0.790 0.0857  4.70   10.2    6.91
#>  5  1995 0.728 0.0855  4.07    8.81   5.99
#>  6  1996 1.20  0.0986 11.4    27.7   17.7 
#>  7  1997 0.835 0.0791  5.37   11.0    7.68
#>  8  1998 1.02  0.0900  7.90   17.8   11.9 
#>  9  1999 1.05  0.143   6.58   23.9   12.5 
#> 10  2000 0.811 0.0989  4.64   11.3    7.25
#> # … with 19 more rows