Get prediction matrix for a fitted GAM

Get prediction matrix for a fitted GAM

Usage

anlz_prdmatrix(mod, doystr = 1, doyend = 364, avemat = FALSE)

Arguments

mod

input model object as returned by anlz_gam

doystr

numeric indicating start Julian day for extracting averages

doyend

numeric indicating ending Julian day for extracting averages

avemat

logical indicating if the prediction matrix is to be passed to anlz_metseason (default) or anlz_avgseason

Value

a data.frame with predictors to use with the fitted GAM

Details

Used internally by anlz_metseason, not to be used by itself

Examples

library(dplyr)

# data to model
tomod <- rawdat %>%
  filter(station %in% 34) %>%
  filter(param %in% 'chl') %>% 
  filter(yr > 2015)
  
mod <- anlz_gam(tomod, trans = 'log10')
anlz_prdmatrix(mod, doystr = 90, doyend = 180)
#>           date   yr doy cont_year
#> 1   2016-03-30 2016  90  2016.243
#> 2   2016-03-31 2016  91  2016.246
#> 3   2016-04-01 2016  92  2016.249
#> 4   2016-04-02 2016  93  2016.251
#> 5   2016-04-03 2016  94  2016.254
#> 6   2016-04-04 2016  95  2016.257
#> 7   2016-04-05 2016  96  2016.260
#> 8   2016-04-06 2016  97  2016.262
#> 9   2016-04-07 2016  98  2016.265
#> 10  2016-04-08 2016  99  2016.268
#> 11  2016-04-09 2016 100  2016.270
#> 12  2016-04-10 2016 101  2016.273
#> 13  2016-04-11 2016 102  2016.276
#> 14  2016-04-12 2016 103  2016.279
#> 15  2016-04-13 2016 104  2016.281
#> 16  2016-04-14 2016 105  2016.284
#> 17  2016-04-15 2016 106  2016.287
#> 18  2016-04-16 2016 107  2016.290
#> 19  2016-04-17 2016 108  2016.292
#> 20  2016-04-18 2016 109  2016.295
#> 21  2016-04-19 2016 110  2016.298
#> 22  2016-04-20 2016 111  2016.301
#> 23  2016-04-21 2016 112  2016.303
#> 24  2016-04-22 2016 113  2016.306
#> 25  2016-04-23 2016 114  2016.309
#> 26  2016-04-24 2016 115  2016.311
#> 27  2016-04-25 2016 116  2016.314
#> 28  2016-04-26 2016 117  2016.317
#> 29  2016-04-27 2016 118  2016.320
#> 30  2016-04-28 2016 119  2016.322
#> 31  2016-04-29 2016 120  2016.325
#> 32  2016-04-30 2016 121  2016.328
#> 33  2016-05-01 2016 122  2016.331
#> 34  2016-05-02 2016 123  2016.333
#> 35  2016-05-03 2016 124  2016.336
#> 36  2016-05-04 2016 125  2016.339
#> 37  2016-05-05 2016 126  2016.342
#> 38  2016-05-06 2016 127  2016.344
#> 39  2016-05-07 2016 128  2016.347
#> 40  2016-05-08 2016 129  2016.350
#> 41  2016-05-09 2016 130  2016.352
#> 42  2016-05-10 2016 131  2016.355
#> 43  2016-05-11 2016 132  2016.358
#> 44  2016-05-12 2016 133  2016.361
#> 45  2016-05-13 2016 134  2016.363
#> 46  2016-05-14 2016 135  2016.366
#> 47  2016-05-15 2016 136  2016.369
#> 48  2016-05-16 2016 137  2016.372
#> 49  2016-05-17 2016 138  2016.374
#> 50  2016-05-18 2016 139  2016.377
#> 51  2016-05-19 2016 140  2016.380
#> 52  2016-05-20 2016 141  2016.383
#> 53  2016-05-21 2016 142  2016.385
#> 54  2016-05-22 2016 143  2016.388
#> 55  2016-05-23 2016 144  2016.391
#> 56  2016-05-24 2016 145  2016.393
#> 57  2016-05-25 2016 146  2016.396
#> 58  2016-05-26 2016 147  2016.399
#> 59  2016-05-27 2016 148  2016.402
#> 60  2016-05-28 2016 149  2016.404
#> 61  2016-05-29 2016 150  2016.407
#> 62  2016-05-30 2016 151  2016.410
#> 63  2016-05-31 2016 152  2016.413
#> 64  2016-06-01 2016 153  2016.415
#> 65  2016-06-02 2016 154  2016.418
#> 66  2016-06-03 2016 155  2016.421
#> 67  2016-06-04 2016 156  2016.423
#> 68  2016-06-05 2016 157  2016.426
#> 69  2016-06-06 2016 158  2016.429
#> 70  2016-06-07 2016 159  2016.432
#> 71  2016-06-08 2016 160  2016.434
#> 72  2016-06-09 2016 161  2016.437
#> 73  2016-06-10 2016 162  2016.440
#> 74  2016-06-11 2016 163  2016.443
#> 75  2016-06-12 2016 164  2016.445
#> 76  2016-06-13 2016 165  2016.448
#> 77  2016-06-14 2016 166  2016.451
#> 78  2016-06-15 2016 167  2016.454
#> 79  2016-06-16 2016 168  2016.456
#> 80  2016-06-17 2016 169  2016.459
#> 81  2016-06-18 2016 170  2016.462
#> 82  2016-06-19 2016 171  2016.464
#> 83  2016-06-20 2016 172  2016.467
#> 84  2016-06-21 2016 173  2016.470
#> 85  2016-06-22 2016 174  2016.473
#> 86  2016-06-23 2016 175  2016.475
#> 87  2016-06-24 2016 176  2016.478
#> 88  2016-06-25 2016 177  2016.481
#> 89  2016-06-26 2016 178  2016.484
#> 90  2016-06-27 2016 179  2016.486
#> 91  2016-06-28 2016 180  2016.489
#> 92  2017-03-31 2017  90  2017.244
#> 93  2017-04-01 2017  91  2017.247
#> 94  2017-04-02 2017  92  2017.249
#> 95  2017-04-03 2017  93  2017.252
#> 96  2017-04-04 2017  94  2017.255
#> 97  2017-04-05 2017  95  2017.258
#> 98  2017-04-06 2017  96  2017.260
#> 99  2017-04-07 2017  97  2017.263
#> 100 2017-04-08 2017  98  2017.266
#> 101 2017-04-09 2017  99  2017.268
#> 102 2017-04-10 2017 100  2017.271
#> 103 2017-04-11 2017 101  2017.274
#> 104 2017-04-12 2017 102  2017.277
#> 105 2017-04-13 2017 103  2017.279
#> 106 2017-04-14 2017 104  2017.282
#> 107 2017-04-15 2017 105  2017.285
#> 108 2017-04-16 2017 106  2017.288
#> 109 2017-04-17 2017 107  2017.290
#> 110 2017-04-18 2017 108  2017.293
#> 111 2017-04-19 2017 109  2017.296
#> 112 2017-04-20 2017 110  2017.299
#> 113 2017-04-21 2017 111  2017.301
#> 114 2017-04-22 2017 112  2017.304
#> 115 2017-04-23 2017 113  2017.307
#> 116 2017-04-24 2017 114  2017.310
#> 117 2017-04-25 2017 115  2017.312
#> 118 2017-04-26 2017 116  2017.315
#> 119 2017-04-27 2017 117  2017.318
#> 120 2017-04-28 2017 118  2017.321
#> 121 2017-04-29 2017 119  2017.323
#> 122 2017-04-30 2017 120  2017.326
#> 123 2017-05-01 2017 121  2017.329
#> 124 2017-05-02 2017 122  2017.332
#> 125 2017-05-03 2017 123  2017.334
#> 126 2017-05-04 2017 124  2017.337
#> 127 2017-05-05 2017 125  2017.340
#> 128 2017-05-06 2017 126  2017.342
#> 129 2017-05-07 2017 127  2017.345
#> 130 2017-05-08 2017 128  2017.348
#> 131 2017-05-09 2017 129  2017.351
#> 132 2017-05-10 2017 130  2017.353
#> 133 2017-05-11 2017 131  2017.356
#> 134 2017-05-12 2017 132  2017.359
#> 135 2017-05-13 2017 133  2017.362
#> 136 2017-05-14 2017 134  2017.364
#> 137 2017-05-15 2017 135  2017.367
#> 138 2017-05-16 2017 136  2017.370
#> 139 2017-05-17 2017 137  2017.373
#> 140 2017-05-18 2017 138  2017.375
#> 141 2017-05-19 2017 139  2017.378
#> 142 2017-05-20 2017 140  2017.381
#> 143 2017-05-21 2017 141  2017.384
#> 144 2017-05-22 2017 142  2017.386
#> 145 2017-05-23 2017 143  2017.389
#> 146 2017-05-24 2017 144  2017.392
#> 147 2017-05-25 2017 145  2017.395
#> 148 2017-05-26 2017 146  2017.397
#> 149 2017-05-27 2017 147  2017.400
#> 150 2017-05-28 2017 148  2017.403
#> 151 2017-05-29 2017 149  2017.405
#> 152 2017-05-30 2017 150  2017.408
#> 153 2017-05-31 2017 151  2017.411
#> 154 2017-06-01 2017 152  2017.414
#> 155 2017-06-02 2017 153  2017.416
#> 156 2017-06-03 2017 154  2017.419
#> 157 2017-06-04 2017 155  2017.422
#> 158 2017-06-05 2017 156  2017.425
#> 159 2017-06-06 2017 157  2017.427
#> 160 2017-06-07 2017 158  2017.430
#> 161 2017-06-08 2017 159  2017.433
#> 162 2017-06-09 2017 160  2017.436
#> 163 2017-06-10 2017 161  2017.438
#> 164 2017-06-11 2017 162  2017.441
#> 165 2017-06-12 2017 163  2017.444
#> 166 2017-06-13 2017 164  2017.447
#> 167 2017-06-14 2017 165  2017.449
#> 168 2017-06-15 2017 166  2017.452
#> 169 2017-06-16 2017 167  2017.455
#> 170 2017-06-17 2017 168  2017.458
#> 171 2017-06-18 2017 169  2017.460
#> 172 2017-06-19 2017 170  2017.463
#> 173 2017-06-20 2017 171  2017.466
#> 174 2017-06-21 2017 172  2017.468
#> 175 2017-06-22 2017 173  2017.471
#> 176 2017-06-23 2017 174  2017.474
#> 177 2017-06-24 2017 175  2017.477
#> 178 2017-06-25 2017 176  2017.479
#> 179 2017-06-26 2017 177  2017.482
#> 180 2017-06-27 2017 178  2017.485
#> 181 2017-06-28 2017 179  2017.488
#> 182 2017-06-29 2017 180  2017.490
#> 183 2018-03-31 2018  90  2018.244
#> 184 2018-04-01 2018  91  2018.247
#> 185 2018-04-02 2018  92  2018.249
#> 186 2018-04-03 2018  93  2018.252
#> 187 2018-04-04 2018  94  2018.255
#> 188 2018-04-05 2018  95  2018.258
#> 189 2018-04-06 2018  96  2018.260
#> 190 2018-04-07 2018  97  2018.263
#> 191 2018-04-08 2018  98  2018.266
#> 192 2018-04-09 2018  99  2018.268
#> 193 2018-04-10 2018 100  2018.271
#> 194 2018-04-11 2018 101  2018.274
#> 195 2018-04-12 2018 102  2018.277
#> 196 2018-04-13 2018 103  2018.279
#> 197 2018-04-14 2018 104  2018.282
#> 198 2018-04-15 2018 105  2018.285
#> 199 2018-04-16 2018 106  2018.288
#> 200 2018-04-17 2018 107  2018.290
#> 201 2018-04-18 2018 108  2018.293
#> 202 2018-04-19 2018 109  2018.296
#> 203 2018-04-20 2018 110  2018.299
#> 204 2018-04-21 2018 111  2018.301
#> 205 2018-04-22 2018 112  2018.304
#> 206 2018-04-23 2018 113  2018.307
#> 207 2018-04-24 2018 114  2018.310
#> 208 2018-04-25 2018 115  2018.312
#> 209 2018-04-26 2018 116  2018.315
#> 210 2018-04-27 2018 117  2018.318
#> 211 2018-04-28 2018 118  2018.321
#> 212 2018-04-29 2018 119  2018.323
#> 213 2018-04-30 2018 120  2018.326
#> 214 2018-05-01 2018 121  2018.329
#> 215 2018-05-02 2018 122  2018.332
#> 216 2018-05-03 2018 123  2018.334
#> 217 2018-05-04 2018 124  2018.337
#> 218 2018-05-05 2018 125  2018.340
#> 219 2018-05-06 2018 126  2018.342
#> 220 2018-05-07 2018 127  2018.345
#> 221 2018-05-08 2018 128  2018.348
#> 222 2018-05-09 2018 129  2018.351
#> 223 2018-05-10 2018 130  2018.353
#> 224 2018-05-11 2018 131  2018.356
#> 225 2018-05-12 2018 132  2018.359
#> 226 2018-05-13 2018 133  2018.362
#> 227 2018-05-14 2018 134  2018.364
#> 228 2018-05-15 2018 135  2018.367
#> 229 2018-05-16 2018 136  2018.370
#> 230 2018-05-17 2018 137  2018.373
#> 231 2018-05-18 2018 138  2018.375
#> 232 2018-05-19 2018 139  2018.378
#> 233 2018-05-20 2018 140  2018.381
#> 234 2018-05-21 2018 141  2018.384
#> 235 2018-05-22 2018 142  2018.386
#> 236 2018-05-23 2018 143  2018.389
#> 237 2018-05-24 2018 144  2018.392
#> 238 2018-05-25 2018 145  2018.395
#> 239 2018-05-26 2018 146  2018.397
#> 240 2018-05-27 2018 147  2018.400
#> 241 2018-05-28 2018 148  2018.403
#> 242 2018-05-29 2018 149  2018.405
#> 243 2018-05-30 2018 150  2018.408
#> 244 2018-05-31 2018 151  2018.411
#> 245 2018-06-01 2018 152  2018.414
#> 246 2018-06-02 2018 153  2018.416
#> 247 2018-06-03 2018 154  2018.419
#> 248 2018-06-04 2018 155  2018.422
#> 249 2018-06-05 2018 156  2018.425
#> 250 2018-06-06 2018 157  2018.427
#> 251 2018-06-07 2018 158  2018.430
#> 252 2018-06-08 2018 159  2018.433
#> 253 2018-06-09 2018 160  2018.436
#> 254 2018-06-10 2018 161  2018.438
#> 255 2018-06-11 2018 162  2018.441
#> 256 2018-06-12 2018 163  2018.444
#> 257 2018-06-13 2018 164  2018.447
#> 258 2018-06-14 2018 165  2018.449
#> 259 2018-06-15 2018 166  2018.452
#> 260 2018-06-16 2018 167  2018.455
#> 261 2018-06-17 2018 168  2018.458
#> 262 2018-06-18 2018 169  2018.460
#> 263 2018-06-19 2018 170  2018.463
#> 264 2018-06-20 2018 171  2018.466
#> 265 2018-06-21 2018 172  2018.468
#> 266 2018-06-22 2018 173  2018.471
#> 267 2018-06-23 2018 174  2018.474
#> 268 2018-06-24 2018 175  2018.477
#> 269 2018-06-25 2018 176  2018.479
#> 270 2018-06-26 2018 177  2018.482
#> 271 2018-06-27 2018 178  2018.485
#> 272 2018-06-28 2018 179  2018.488
#> 273 2018-06-29 2018 180  2018.490
#> 274 2019-03-31 2019  90  2019.244
#> 275 2019-04-01 2019  91  2019.247
#> 276 2019-04-02 2019  92  2019.249
#> 277 2019-04-03 2019  93  2019.252
#> 278 2019-04-04 2019  94  2019.255
#> 279 2019-04-05 2019  95  2019.258
#> 280 2019-04-06 2019  96  2019.260
#> 281 2019-04-07 2019  97  2019.263
#> 282 2019-04-08 2019  98  2019.266
#> 283 2019-04-09 2019  99  2019.268
#> 284 2019-04-10 2019 100  2019.271
#> 285 2019-04-11 2019 101  2019.274
#> 286 2019-04-12 2019 102  2019.277
#> 287 2019-04-13 2019 103  2019.279
#> 288 2019-04-14 2019 104  2019.282
#> 289 2019-04-15 2019 105  2019.285
#> 290 2019-04-16 2019 106  2019.288
#> 291 2019-04-17 2019 107  2019.290
#> 292 2019-04-18 2019 108  2019.293
#> 293 2019-04-19 2019 109  2019.296
#> 294 2019-04-20 2019 110  2019.299
#> 295 2019-04-21 2019 111  2019.301
#> 296 2019-04-22 2019 112  2019.304
#> 297 2019-04-23 2019 113  2019.307
#> 298 2019-04-24 2019 114  2019.310
#> 299 2019-04-25 2019 115  2019.312
#> 300 2019-04-26 2019 116  2019.315
#> 301 2019-04-27 2019 117  2019.318
#> 302 2019-04-28 2019 118  2019.321
#> 303 2019-04-29 2019 119  2019.323
#> 304 2019-04-30 2019 120  2019.326
#> 305 2019-05-01 2019 121  2019.329
#> 306 2019-05-02 2019 122  2019.332
#> 307 2019-05-03 2019 123  2019.334
#> 308 2019-05-04 2019 124  2019.337
#> 309 2019-05-05 2019 125  2019.340
#> 310 2019-05-06 2019 126  2019.342
#> 311 2019-05-07 2019 127  2019.345
#> 312 2019-05-08 2019 128  2019.348
#> 313 2019-05-09 2019 129  2019.351
#> 314 2019-05-10 2019 130  2019.353
#> 315 2019-05-11 2019 131  2019.356
#> 316 2019-05-12 2019 132  2019.359
#> 317 2019-05-13 2019 133  2019.362
#> 318 2019-05-14 2019 134  2019.364
#> 319 2019-05-15 2019 135  2019.367
#> 320 2019-05-16 2019 136  2019.370
#> 321 2019-05-17 2019 137  2019.373
#> 322 2019-05-18 2019 138  2019.375
#> 323 2019-05-19 2019 139  2019.378
#> 324 2019-05-20 2019 140  2019.381
#> 325 2019-05-21 2019 141  2019.384
#> 326 2019-05-22 2019 142  2019.386
#> 327 2019-05-23 2019 143  2019.389
#> 328 2019-05-24 2019 144  2019.392
#> 329 2019-05-25 2019 145  2019.395
#> 330 2019-05-26 2019 146  2019.397
#> 331 2019-05-27 2019 147  2019.400
#> 332 2019-05-28 2019 148  2019.403
#> 333 2019-05-29 2019 149  2019.405
#> 334 2019-05-30 2019 150  2019.408
#> 335 2019-05-31 2019 151  2019.411
#> 336 2019-06-01 2019 152  2019.414
#> 337 2019-06-02 2019 153  2019.416
#> 338 2019-06-03 2019 154  2019.419
#> 339 2019-06-04 2019 155  2019.422
#> 340 2019-06-05 2019 156  2019.425
#> 341 2019-06-06 2019 157  2019.427
#> 342 2019-06-07 2019 158  2019.430
#> 343 2019-06-08 2019 159  2019.433
#> 344 2019-06-09 2019 160  2019.436
#> 345 2019-06-10 2019 161  2019.438
#> 346 2019-06-11 2019 162  2019.441
#> 347 2019-06-12 2019 163  2019.444
#> 348 2019-06-13 2019 164  2019.447
#> 349 2019-06-14 2019 165  2019.449
#> 350 2019-06-15 2019 166  2019.452
#> 351 2019-06-16 2019 167  2019.455
#> 352 2019-06-17 2019 168  2019.458
#> 353 2019-06-18 2019 169  2019.460
#> 354 2019-06-19 2019 170  2019.463
#> 355 2019-06-20 2019 171  2019.466
#> 356 2019-06-21 2019 172  2019.468
#> 357 2019-06-22 2019 173  2019.471
#> 358 2019-06-23 2019 174  2019.474
#> 359 2019-06-24 2019 175  2019.477
#> 360 2019-06-25 2019 176  2019.479
#> 361 2019-06-26 2019 177  2019.482
#> 362 2019-06-27 2019 178  2019.485
#> 363 2019-06-28 2019 179  2019.488
#> 364 2019-06-29 2019 180  2019.490