Estimate seasonal rates of change based on average estimates

Estimate seasonal rates of change based on average estimates

anlz_trndseason(
  moddat = NULL,
  mods = NULL,
  doystr = 1,
  doyend = 364,
  justify = c("left", "right", "center"),
  win = 5,
  ...
)

Arguments

moddat
input raw data, one station and paramater

mods
optional list of model objects

doystr
numeric indicating start Julian day for extracting averages

doyend
numeric indicating ending Julian day for extracting averages

justify
chr string indicating the justification for the trend window

win
numeric indicating number of years to use for the trend window

...
additional arguments passed to other methods

Value

A data frame of slope estimates and p-values for each year

Details

Trends are based on the slope of the fitted linear trend within the window, where the linear trend is estimated using a meta-analysis regression model (from anlz_mixmeta) for the seasonal averages (from anlz_avgseason)

See also

Examples

library(dplyr)

# fit models with function
tomod <- rawdat %>%
  filter(station %in% 32) %>%
  filter(param %in% 'chl')
if (FALSE) {
anlz_trndseason(tomod, trans = 'log10', doystr = 90, doyend = 180, justify = 'left', win = 5)
}
# use previously fitted list of models
trans <- 'log10'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
anlz_trndseason(mods = mods, trans = 'log10', doystr = 90, doyend = 180, justify = 'left', win = 5)

#> # A tibble: 90 x 9
#> # Groups:   model [3]
#>    model   avg     se    yr bt_avg bt_lwr bt_upr  yrcoef  pval
#>    <chr> <dbl>  <dbl> <int>  <dbl>  <dbl>  <dbl>   <dbl> <dbl>
#>  1 gam0  0.776 0.0312  1990   7.60   6.60   8.75 0.00493 0.595
#>  2 gam0  0.781 0.0302  1991   7.69   6.71   8.81 0.00493 0.583
#>  3 gam0  0.786 0.0292  1992   7.77   6.81   8.87 0.00493 0.571
#>  4 gam0  0.790 0.0283  1993   7.86   6.92   8.93 0.00493 0.560
#>  5 gam0  0.795 0.0275  1994   7.95   7.02   9.00 0.00493 0.549
#>  6 gam0  0.800 0.0266  1995   8.04   7.13   9.07 0.00493 0.538
#>  7 gam0  0.805 0.0259  1996   8.13   7.24   9.14 0.00493 0.528
#>  8 gam0  0.810 0.0252  1997   8.23   7.34   9.22 0.00493 0.519
#>  9 gam0  0.815 0.0246  1998   8.32   7.45   9.30 0.00493 0.511
#> 10 gam0  0.820 0.0240  1999   8.42   7.55   9.38 0.00493 0.504
#> # … with 80 more rows