Estimate seasonal rates of change based on average estimates

Estimate seasonal rates of change based on average estimates

anlz_trndseason(
  moddat = NULL,
  mods = NULL,
  doystr = 1,
  doyend = 364,
  justify = c("left", "right", "center"),
  win = 5,
  ...
)

Arguments

moddat
input raw data, one station and paramater

mods
optional list of model objects

doystr
numeric indicating start Julian day for extracting averages

doyend
numeric indicating ending Julian day for extracting averages

justify
chr string indicating the justification for the trend window

win
numeric indicating number of years to use for the trend window

...
additional arguments passed to other methods

Value

A data frame of slope estimates and p-values for each year

Details

Trends are based on the slope of the fitted linear trend within the window, where the linear trend is estimated using a meta-analysis regression model (from anlz_mixmeta) for the seasonal averages (from anlz_avgseason)

See also

Examples

library(dplyr)

# fit models with function
tomod <- rawdat %>%
  filter(station %in% 32) %>%
  filter(param %in% 'chl')
if (FALSE) {
anlz_trndseason(tomod, trans = 'boxcox', doystr = 90, doyend = 180, justify = 'left', win = 5)
}
# use previously fitted list of models
trans <- 'boxcox'
mods <- list(
  gam0 = anlz_gam(tomod, mod = 'gam0', trans = trans),
  gam1 = anlz_gam(tomod, mod = 'gam1', trans = trans), 
  gam2 = anlz_gam(tomod, mod = 'gam2', trans = trans)
  )
anlz_trndseason(mods = mods, trans = 'boxcox', doystr = 90, doyend = 180, justify = 'left', win = 5)

#> # A tibble: 84 x 6
#> # Groups:   model [3]
#>    model predicted    se    yr yrcoef  pval
#>    <chr>     <dbl> <dbl> <int>  <dbl> <dbl>
#>  1 gam0       8.00  1.05  1990  0.181 0.584
#>  2 gam0       8.17  1.05  1991  0.186 0.573
#>  3 gam0       8.35  1.05  1992  0.191 0.562
#>  4 gam0       8.53  1.04  1993  0.196 0.551
#>  5 gam0       8.72  1.04  1994  0.202 0.540
#>  6 gam0       8.92  1.04  1995  0.207 0.529
#>  7 gam0       9.11  1.04  1996  0.213 0.517
#>  8 gam0       9.32  1.04  1997  0.219 0.505
#>  9 gam0       9.53  1.04  1998  0.225 0.493
#> 10 gam0       9.74  1.04  1999  0.231 0.481
#> # … with 74 more rows