Estimate seasonal rates of change based on average estimates

Estimate seasonal rates of change based on average estimates

anlz_trndseason(
  mod,
  metfun = mean,
  doystr = 1,
  doyend = 364,
  justify = c("center", "left", "right"),
  win = 5,
  nsim = 10000,
  useave = FALSE,
  ...
)

Arguments

mod

input model object as returned by anlz_gam

metfun

function input for metric to calculate, e.g., mean, var, max, etc

doystr

numeric indicating start Julian day for extracting averages

doyend

numeric indicating ending Julian day for extracting averages

justify

chr string indicating the justification for the trend window

win

numeric indicating number of years to use for the trend window, see details

nsim

numeric indicating number of random draws for simulating uncertainty

useave

logical indicating if anlz_avgseason is used for the seasonal metric calculation

...

additional arguments passed to metfun, e.g., na.rm = TRUE)

Value

A data frame of slope estimates and p-values for each year

Details

Trends are based on the slope of the fitted linear trend within the window, where the linear trend is estimated using a meta-analysis regression model (from anlz_mixmeta) for the seasonal metrics (from anlz_metseason).

Note that for left and right windows, the exact number of years in win is used. For example, a left-centered window for 1990 of ten years will include exactly ten years from 1990, 1991, ... , 1999. The same applies to a right-centered window, e.g., for 1990 it would include 1981, 1982, ..., 1990 (if those years have data). However, for a centered window, picking an even number of years for the window width will create a slightly off-centered window because it is impossible to center on an even number of years. For example, if win = 8 and justify = 'center', the estimate for 2000 will be centered on 1997 to 2004 (three years left, four years right, eight years total). Centering for window widths with an odd number of years will always create a symmetrical window, i.e., if win = 7 and justify = 'center', the estimate for 2000 will be centered on 1997 and 2003 (three years left, three years right, seven years total).

See also

Examples

library(dplyr)

# data to model
tomod <- rawdat %>%
  filter(station %in% 34) %>%
  filter(param %in% 'chl')

mod <- anlz_gam(tomod, trans = 'log10')
anlz_trndseason(mod, doystr = 90, doyend = 180, justify = 'center', win = 8)
#> # A tibble: 29 × 11
#>       yr   met     se bt_lwr bt_upr bt_met    yrcoef     pval appr_yrcoef
#>    <dbl> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>     <dbl>    <dbl>       <dbl>
#>  1  1991 0.113 0.535   0.130  16.3    1.46 NaN       NaN          NA     
#>  2  1992 0.684 0.0655  4.03    7.28   5.41 NaN       NaN          NA     
#>  3  1993 0.689 0.0985  3.51    8.55   5.48 NaN       NaN          NA     
#>  4  1994 0.790 0.0864  4.68   10.2    6.91   0.0655    0.0151      1.14  
#>  5  1995 0.728 0.0799  4.18    8.59   5.99   0.0584    0.0102      1.16  
#>  6  1996 1.20  0.0911 11.8    26.8   17.7    0.0309    0.268       0.623 
#>  7  1997 0.835 0.0906  5.10   11.6    7.68   0.00456   0.872       0.0940
#>  8  1998 1.02  0.0790  8.30   16.9   11.9   -0.00388   0.887      -0.0837
#>  9  1999 1.05  0.154   6.25   25.1   12.5   -0.0383    0.0497     -0.847 
#> 10  2000 0.811 0.0977  4.67   11.3    7.25  -0.0170    0.263      -0.335 
#> # … with 19 more rows, and 2 more variables: yrcoef_lwr <dbl>, yrcoef_upr <dbl>