Estimate seasonal rates of change based on average estimates

Estimate seasonal rates of change based on average estimates

anlz_trndseason(
  mod,
  doystr = 1,
  doyend = 364,
  justify = c("center", "left", "right"),
  win = 5
)

Arguments

mod
input model object as returned by anlz_gam

doystr
numeric indicating start Julian day for extracting averages

doyend
numeric indicating ending Julian day for extracting averages

justify
chr string indicating the justification for the trend window

win
numeric indicating number of years to use for the trend window

Value

A data frame of slope estimates and p-values for each year

Details

Trends are based on the slope of the fitted linear trend within the window, where the linear trend is estimated using a meta-analysis regression model (from anlz_mixmeta) for the seasonal averages (from anlz_avgseason).

Note that for left and right windows, the exact number of years in win is used. For example, a left-centered window for 1990 of ten years will include exactly ten years from 1990, 1991, ... , 1999. The same applies to a right-centered window, e.g., for 1990 it would include 1981, 1982, ..., 1990 (if those years have data). However, for a centered window, picking an even number of years will always have an extra year to center the window exactly, e.g., a ten year window for 1990 will include eleven years from 1985 to 1995 so there is the same number of years to the left and right of center. A centered window with an odd number of years will always be centered and includes the exact number of years used in win.

See also

Examples

library(dplyr)

# data to model
tomod <- rawdat %>%
  filter(station %in% 34) %>%
  filter(param %in% 'chl')

mod <- anlz_gam(tomod, trans = 'log10')
anlz_trndseason(mod, doystr = 90, doyend = 180, justify = 'center', win = 8)

#> # A tibble: 29 x 11
#>      avg     se    yr bt_lwr bt_upr bt_avg    yrcoef      pval appr_yrcoef
#>    <dbl>  <dbl> <int>  <dbl>  <dbl>  <dbl>     <dbl>     <dbl>       <dbl>
#>  1 0.116 0.531   1991  0.133  16.1    1.46 NaN       NaN            NA    
#>  2 0.684 0.0658  1992  4.02    7.29   5.41 NaN       NaN            NA    
#>  3 0.692 0.0964  1993  3.57    8.53   5.52 NaN       NaN            NA    
#>  4 0.791 0.0861  1994  4.70   10.2    6.93 NaN       NaN            NA    
#>  5 0.731 0.0804  1995  4.20    8.67   6.03   0.0624    0.00470       1.17 
#>  6 1.20  0.0914  1996 11.7    26.8   17.7    0.0362    0.0863        0.697
#>  7 0.839 0.0903  1997  5.15   11.6    7.74   0.0174    0.455         0.344
#>  8 1.03  0.0794  1998  8.37   17.1   12.0    0.00639   0.770         0.134
#>  9 1.05  0.154   1999  6.32   25.3   12.7   -0.0123    0.564        -0.256
#> 10 0.811 0.0984  2000  4.65   11.3    7.25  -0.0332    0.0293       -0.722
#> # … with 19 more rows, and 2 more variables: yrcoef_lwr <dbl>, yrcoef_upr <dbl>