Estimate seasonal rates of change based on average estimates

Estimate seasonal rates of change based on average estimates

anlz_trndseason(
  mod,
  metfun = mean,
  doystr = 1,
  doyend = 364,
  justify = c("center", "left", "right"),
  win = 5,
  nsim = 10000,
  ...
)

Arguments

mod
input model object as returned by anlz_gam

metfun
function input for metric to calculate, e.g., mean, var, max, etc

doystr
numeric indicating start Julian day for extracting averages

doyend
numeric indicating ending Julian day for extracting averages

justify
chr string indicating the justification for the trend window

win
numeric indicating number of years to use for the trend window

nsim
numeric indicating number of random draws for simulating uncertainty

...
additional arguments passed to metfun, e.g., na.rm = TRUE)

Value

A data frame of slope estimates and p-values for each year

Details

Trends are based on the slope of the fitted linear trend within the window, where the linear trend is estimated using a meta-analysis regression model (from anlz_mixmeta) for the seasonal metrics (from anlz_metseason).

Note that for left and right windows, the exact number of years in win is used. For example, a left-centered window for 1990 of ten years will include exactly ten years from 1990, 1991, ... , 1999. The same applies to a right-centered window, e.g., for 1990 it would include 1981, 1982, ..., 1990 (if those years have data). However, for a centered window, picking an even number of years will always have an extra year to center the window exactly, e.g., a ten year window for 1990 will include eleven years from 1985 to 1995 so there is the same number of years to the left and right of center. A centered window with an odd number of years will always be centered and includes the exact number of years used in win.

See also

Examples

library(dplyr)

# data to model
tomod <- rawdat %>%
  filter(station %in% 34) %>%
  filter(param %in% 'chl')

mod <- anlz_gam(tomod, trans = 'log10')
anlz_trndseason(mod, doystr = 90, doyend = 180, justify = 'center', win = 8)

#> # A tibble: 29 × 11
#>       yr   met     se bt_lwr bt_upr bt_met    yrcoef      pval appr_yrcoef
#>    <dbl> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>     <dbl>     <dbl>       <dbl>
#>  1  1991 0.113 0.531   0.133  16.0    1.46 NaN       NaN            NA    
#>  2  1992 0.684 0.0657  4.02    7.28   5.41 NaN       NaN            NA    
#>  3  1993 0.689 0.0970  3.54    8.49   5.48 NaN       NaN            NA    
#>  4  1994 0.790 0.0858  4.69   10.2    6.91 NaN       NaN            NA    
#>  5  1995 0.728 0.0809  4.16    8.63   5.99   0.0619    0.00541       1.15 
#>  6  1996 1.20  0.0917 11.7    26.8   17.7    0.0360    0.0877        0.690
#>  7  1997 0.835 0.0893  5.13   11.5    7.68   0.0176    0.451         0.345
#>  8  1998 1.02  0.0801  8.26   17.0   11.9    0.00647   0.767         0.135
#>  9  1999 1.05  0.154   6.26   25.1   12.5   -0.0122    0.569        -0.252
#> 10  2000 0.811 0.0989  4.64   11.3    7.25  -0.0329    0.0322       -0.711
#> # … with 19 more rows, and 2 more variables: yrcoef_lwr <dbl>, yrcoef_upr <dbl>